feature: 게임 생성 설정 엔티티에 할당
Former-commit-id: ad8a4ed76cee994403d293bdd4c1b94bdee12196

+++ back/src/main/java/bunsan/returnz/domain/game/dto/RequestSettingGame.java
@@ -29,7 +29,7 @@ public class RequestSettingGame {
 	private LocalDateTime startTime;
 	@Nullable
 	private Integer totalTurn;
-	private List<Integer> memberIdList;
+	private List<Long> memberIdList;
 
 	public void validateRequestSettingGame() {
